work on test coverage
